unit_azure_tests/requirements.txt
azure-cognitiveservices-speech
pytest
